[sql lab] preserve schema on visualize flow
Also major refactoring:
* moving most SQL Lab views into views/sqllab.py
* corresponding redirect for backward compatibility
* new unit test for the backend part of the visualize flow
* fixing objects that were double-serialized to json for no good reason
