+++ UPDATING.MD
@@ -3,6 +3,10 @@
 This file documents any backwards-incompatible changes in Superset and
 assists people when migrating to a new version.
 
+## Superset 0.25.0
+
+* [4565](https://github.com/apache/incubator-superset/pull/4565)
+
 ## Superset 0.23.0
 
 * [4565](https://github.com/apache/incubator-superset/pull/4565)
