+++ superset/assets/javascripts/SqlLab/actions.js
@@ -409,8 +409,7 @@ export function createDatasourceStarted() {
   return { type: CREATE_DATASOURCE_STARTED };
 }
 export function createDatasourceSuccess(response) {
-  const data = JSON.parse(response);
-  const datasource = `${data.table_id}__table`;
+  const datasource = `${response.table_id}__table`;
   return { type: CREATE_DATASOURCE_SUCCESS, datasource };
 }
 export function createDatasourceFailed(err) {
@@ -423,7 +422,7 @@ export function createDatasource(vizOptions, context) {
 
     return $.ajax({
       type: 'POST',
-      url: '/superset/sqllab_viz/',
+      url: '/sqllab/sqllab_viz/',
       async: false,
       data: {
         data: JSON.stringify(vizOptions),
